feat(dashboard): the slide appears in the delete dialog itself
